gen-manifests: separate worker queue code
Add the worker queue code to a separate file for better organisation
and readability.
